proyecto-rag-cepet/requirements.txt
# --- IA y LangChain ---
langchain
langchain-core
langchain-openai
langchain-postgres
langchain-text-splitters
langchain-community
langchain-unstructured

# --- BBDD ---
psycopg2-binary
pgvector
#SQLAlchemy

# --- Carga de Documentos ---
unstructured[local-inference]
pdf2image
pdfminer.six

# --- Frontend ---
streamlit

# --- Utilidades ---
python-dotenv